Refactoring
Moved all command codes to the top of the classes for a better overview
of command codes and easier changing in case of bugs. Also makes the
commands a bit more 'human readable'.

Added a new method in the 'RGBWLights' class for setting a color and
renamed the 'SetColor' method to 'SetTrueColor'.

Changed some of the access modifiers and made certain properties
publicly 'getable'.

Made some small edits in documentation.